A persistent stash commit for testing enable_basic_auth
It appears that our local_testing inventory is for a 3 host config,
likely managed by Terraform. However, the 2018 team is currently not
using Terraform and therefore I just use this to test that
enable_basic_auth works as expected in a `vagrant up`---most
importantly, the nginx config should allow the shell server to connect
and no one else except with htpasswd.